% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{Aggregate_data}
\alias{Aggregate_data}
\title{Create aggregated data for DIRECT-NET}
\usage{
Aggregate_data(
  object,
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction.name = NULL,
  size_factor_normalize = TRUE,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{k_neigh}{Number of cells to be aggregated per group.}

\item{atacbinary}{Logical, whether the aggregated scATAC-seq data need binary}

\item{max_overlap}{The maximum overlapping ratio of two groups.}

\item{reduction.name}{The reduction name of extracting the cell coordinates used for aggregating.}

\item{size_factor_normalize}{Logical, should accessibility values be normalized by size factor}

\item{seed}{Random seed}

\item{verbose}{Logical, should warning and info messages be printed?}
}
\value{
Aggregated Seurat object.
}
\description{
Function to generate aggregated inputs for DIRECT-NET. \code{Aggregate_data}
}
