% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{asSparseMatrix}
\alias{asSparseMatrix}
\title{Convert a slam matrix to a sparseMatrix}
\usage{
asSparseMatrix(simpleTripletMatrix)
}
\arguments{
\item{simpleTripletMatrix}{A slam matrix}
}
\description{
Convert a slam matrix to a sparseMatrix
}
