% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_CRE}
\alias{generate_CRE}
\title{extract CREs of markers}
\usage{
generate_CRE(L_G_record, P_L_G_record, da_peaks_list)
}
\arguments{
\item{L_G_record}{a list of CRE-Gene relationships}

\item{P_L_G_record}{a list of Promoter-Gene relationships}

\item{da_peaks_list}{a list of DA of each group}
}
\description{
extract CREs of markers
}
