% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_CRE_Gene_links}
\alias{generate_CRE_Gene_links}
\title{extract CREs-gene relations of markers}
\usage{
generate_CRE_Gene_links(direct.net_result, markers)
}
\arguments{
\item{direct.net_result}{dataframe of the result of DIRECT-NET}

\item{markers}{two column dataframe data with marker genes and group information}
}
\description{
extract CREs-gene relations of markers
}
