% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_links_for_Cytoscape}
\alias{generate_links_for_Cytoscape}
\title{generate links for Cytoscape}
\usage{
generate_links_for_Cytoscape(
  L_G_record,
  L_TF_record,
  P_L_G_record,
  P_L_TF_record,
  groups
)
}
\arguments{
\item{L_G_record}{A list of CRE-gene}

\item{L_TF_record}{A list of CRE-TF}

\item{P_L_G_record}{A list of CRE-gene about promoters}

\item{P_L_TF_record}{A list of CRE-TF about promoters}

\item{groups}{cell type information}
}
\value{
TF-G dataframe
}
\description{
generate links for Cytoscape
}
