% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_node_for_Cytoscape}
\alias{generate_node_for_Cytoscape}
\title{generate Node character for Cytoscape}
\usage{
generate_node_for_Cytoscape(network_links, markers)
}
\arguments{
\item{network_links}{Links obtained by generate_links_for_Cytoscape}

\item{markers}{Dataframe with marker and cell type information}
}
\value{
Node character dataframe
}
\description{
generate Node character for Cytoscape
}
