% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_peak_TF_links}
\alias{generate_peak_TF_links}
\title{Identify TFs enriched in CREs of focus markers}
\usage{
generate_peak_TF_links(peaks_bed_list, species, genome, markers)
}
\arguments{
\item{peaks_bed_list}{A list of peaks bed file of each group}

\item{species}{Species used to detect TF}

\item{genome}{For example: BSgenome.Hsapiens.UCSC.hg19}

\item{markers}{two column dataframe data with marker genes and group information}
}
\description{
Identify TFs enriched in CREs of focus markers
}
