% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_TF_G}
\alias{get_TF_G}
\title{Obtain TF and Gene relationships}
\usage{
get_TF_G(L_G_record, L_TF_record, groups)
}
\arguments{
\item{L_G_record}{A list of CRE-gene}

\item{L_TF_record}{A list of CRE-TF}

\item{groups}{cluster information}
}
\value{
TF-G dataframe
}
\description{
Obtain TF and Gene relationships
}
