% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{isSparseMatrix}
\alias{isSparseMatrix}
\title{Convert a sparseMatrix from Matrix package to a slam matrix}
\usage{
isSparseMatrix(x)
}
\arguments{
\item{x}{The sparseMatrix data}
}
\description{
Convert a sparseMatrix from Matrix package to a slam matrix
}
