% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Plot_connections}
\alias{Plot_connections}
\title{Plot links of DIRECT-NET on focused genes}
\usage{
Plot_connections(
  DIRECT_NET_Result,
  gene_anno,
  marker,
  cutoff = NULL,
  upstream = 250000,
  downstream = 250000
)
}
\arguments{
\item{DIRECT_NET_Result}{DIRECT-NET result}

\item{gene_anno}{the genome annotation}

\item{marker}{the focused gene}

\item{cutoff}{the cutoff of importance scores, CREs are links with importance scores higher than cutoff}

\item{upstream}{the number of distance upstream to TSS}

\item{downstream}{the number of distance downstream to TSS}
}
\description{
Plot links of DIRECT-NET on focused genes
}
