% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{Run_DIRECT_NET}
\alias{Run_DIRECT_NET}
\title{run DIRECT-NET on Seurat object}
\usage{
Run_DIRECT_NET(
  object,
  peakcalling = FALSE,
  macs2.path = NULL,
  fragments = NULL,
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction.name = NULL,
  size_factor_normalize = FALSE,
  genome.info,
  focus_markers,
  params = NULL,
  nthread = 2,
  early_stop = FALSE,
  HC_cutoff = NULL,
  LC_cutoff = NULL,
  rescued = FALSE,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{peakcalling}{call peak}

\item{macs2.path}{path to macs2}

\item{fragments}{fragments file}

\item{k_neigh}{Number of cells to aggregate per group.}

\item{atacbinary}{Logical, should accessibility values be binarized}

\item{max_overlap}{The maximum overlapping ratio of two groups.}

\item{reduction.name}{The reduction name of extracting the cell coordinates used for aggregating.}

\item{size_factor_normalize}{Logical, whether need to do size normalization}

\item{genome.info}{the TSS information of genome, e.g. hg19, hg38}

\item{focus_markers}{the focused genes}

\item{params}{the list of parameters used in Xgboost}

\item{nthread}{the number of threads can be manually specified in Xgboost trainning stage, default is 2}

\item{early_stop}{Logical, whether use early stop rule on validation data to reduce overfitting}

\item{HC_cutoff}{the threshold of high functional CREs}

\item{LC_cutoff}{the threshold of low functional CREs}

\item{rescued}{Logical, whether to rescue highly correlated CREs}

\item{seed}{Random seed}

\item{verbose}{Logical, should warning and info messages be printed}
}
\value{
a Seurat object with new links assay.
}
\description{
run DIRECT-NET on Seurat object
}
